[fix][cli] Dedicated docs tools
