+++ pom.xml
@@ -869,14 +869,12 @@ flexible messaging model and an intuitive client API.</description>
         <groupId>io.swagger</groupId>
         <artifactId>swagger-core</artifactId>
         <version>${swagger.version}</version>
-        <scope>provided</scope>
       </dependency>
 
       <dependency>
         <groupId>io.swagger</groupId>
         <artifactId>swagger-annotations</artifactId>
         <version>${swagger.version}</version>
-        <scope>provided</scope>
       </dependency>
 
       <dependency>
@@ -2221,6 +2219,7 @@ flexible messaging model and an intuitive client API.</description>
         <module>pulsar-client-tools-customcommand-example</module>
         <module>pulsar-client-tools-test</module>
         <module>pulsar-client-all</module>
+        <module>pulsar-docs-tools</module>
         <module>pulsar-websocket</module>
         <module>pulsar-proxy</module>
         <module>pulsar-testclient</module>
@@ -2283,6 +2282,7 @@ flexible messaging model and an intuitive client API.</description>
         <module>pulsar-client-tools</module>
         <module>pulsar-client-tools-customcommand-example</module>
         <module>pulsar-client-tools-test</module>
+        <module>pulsar-docs-tools</module>
         <module>pulsar-websocket</module>
         <module>pulsar-proxy</module>
         <module>pulsar-testclient</module>
