+++ pulsar-client-tools/pom.xml
@@ -107,6 +107,7 @@
     <dependency>
       <groupId>io.swagger</groupId>
       <artifactId>swagger-core</artifactId>
+      <scope>provided</scope>
     </dependency>
 
     <!-- functions related dependencies (begin) -->
