+++ pulsar-docs-tools/src/main/java/org/apache/pulsar/docs/tools/package-info.java
@@ -16,4 +16,4 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.pulsar.proxy.util;
+package org.apache.pulsar.docs.tools;
