+++ pulsar-docs-tools/src/test/java/org/apache/pulsar/docs/tools/CmdGenerateDocumentationTest.java
@@ -16,21 +16,36 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.pulsar.proxy.util;
 
-import static org.testng.Assert.assertTrue;
+package org.apache.pulsar.docs.tools;
+
 import java.io.ByteArrayOutputStream;
 import java.io.PrintStream;
 import java.lang.reflect.Field;
 import org.apache.pulsar.common.configuration.FieldContext;
 import org.testng.annotations.Test;
+import static org.testng.Assert.assertTrue;
 
-public class CmdTest {
-
+public class CmdGenerateDocumentationTest {
     @Test
     public void cmdParserProxyConfigurationTest() throws Exception {
         String value = generateDoc("org.apache.pulsar.proxy.server.ProxyConfiguration");
         assertTrue(value.contains("Pulsar proxy"));
+    }
+
+    @Test
+    public void cmdParserTest() throws Exception {
+        generateDoc("org.apache.pulsar.broker.ServiceConfiguration");
+    }
+
+    @Test
+    public void cmdParserWebSocketTest() throws Exception {
+        generateDoc("org.apache.pulsar.websocket.service.WebSocketProxyConfiguration");
+    }
+
+    @Test
+    public void cmdParserClientTest() throws Exception {
+        generateDoc("org.apache.pulsar.client.impl.conf.ClientConfigurationData");
     }
 
     private String generateDoc(String clazz) throws Exception {
@@ -40,7 +55,7 @@ public class CmdTest {
             System.setOut(cacheStream);
             CmdGenerateDocumentation.main(("-c " + clazz).split(" "));
             String message = baoStream.toString();
-            Class cls = Class.forName(clazz);
+            Class<?> cls = Class.forName(clazz);
             Field[] fields = cls.getDeclaredFields();
             for (Field field : fields) {
                 field.setAccessible(true);
